Test commit

+++ src/main/java/web/dao/RoleDao.java
@@ -7,8 +7,12 @@ import java.util.List;
 
 public interface RoleDao {
     List<Role> getRoles();
+
     void save(Role role);
+
     void update(Role role);
-    void delete(int id);
+
+    void delete(long id);
+
     Role getRoleByName(String name);
 }
