+++ src/main/java/web/service/RoleService.java
@@ -7,8 +7,12 @@ import java.util.List;
 
 public interface RoleService {
     List<Role> getRoles();
+
     void save(Role role);
+
     void update(Role role);
-    void delete(int id);
+
+    void delete(long id);
+
     Role getRoleByName(String name);
 }
